Macro fixes
Added an int cast before the unsigned short cast. If the input to these functions is a float (which it is in the camera yaw functions, then casting directly to unsigned short can produce incorrect values when the float is negative.

+++ platform_tilt_brute_forcer/platform_tilt_brute_forcer/FST.cpp
@@ -5,9 +5,9 @@
 
 # define M_PI                  3.14159265358979323846  /* pi */
 
-# define sinsG(x)              gSineTableG[(unsigned short) (x) >> 4]
-# define cossG(x)              gCosineTableG[(unsigned short) (x) >> 4]
-# define revAtansG(x)          gReverseArctanTableG[(unsigned short) (x)]
+# define sinsG(x)              gSineTableG[(unsigned short)(int) (x) >> 4]
+# define cossG(x)              gCosineTableG[(unsigned short)(int) (x) >> 4]
+# define revAtansG(x)          gReverseArctanTableG[(unsigned short)(int) (x)]
 # define sign(x)               (((x) > 0) - ((x) < 0))
 
 class SurfaceG {
